+++ lingochat/src/App.js
@@ -1,9 +1,10 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Home from './Home';
+import LoggedIn from './LoggedIn';
 import Chat from './Chat';
 import IO from 'socket.io-client';
 import './App.css';
 import FindChatRoom from './FindChatRoom';
+import Login from './Login';
 const socket = IO.connect('http://localhost:4000');
 
 function App() {
@@ -11,9 +12,10 @@ function App() {
     <BrowserRouter>
       <div>
         <Routes>
-          <Route path="/" element={<Home socket={socket} />}></Route>
+          <Route path="/loggedIn" element={<LoggedIn socket={socket} />}></Route>
           <Route path="/FindRoom" element={<FindChatRoom socket={socket} />}></Route>
           <Route path="/chat" element={<Chat socket={socket} />}></Route>
+          <Route path="/" element={<Login socket={socket} />}></Route>
         </Routes>
       </div>
     </BrowserRouter>
